RF_12.000.01 | Combining Footer/REST API related classes in one class
